Moved doc prettification intention and all php docs prettifiers to the `intentions` package
