+++ src/main/kotlin/com/vk/kphpstorm/inspections/quickfixes/PrettifySomethingInDocBlockQuickFix.kt
@@ -3,7 +3,7 @@ package com.vk.kphpstorm.inspections.quickfixes
 import com.intellij.codeInspection.LocalQuickFix
 import com.intellij.codeInspection.ProblemDescriptor
 import com.intellij.openapi.project.Project
-import com.vk.kphpstorm.inspections.prettifier.PhpDocPrettification
+import com.vk.kphpstorm.intentions.prettifier.PhpDocPrettification
 
 class PrettifySomethingInDocBlockQuickFix(private val prettification: PhpDocPrettification) : LocalQuickFix {
